+++ tests/BuildResponseTest.php
@@ -127,7 +127,7 @@ class BuildResponseTest extends Base\ResponseBuilderTestCaseBase
 
 		$message_or_api_code = [];    // invalid
 
-		$this->callMakeMethod($api_codes_class_name::OK, $message_or_api_code);
+		$this->callMakeMethod(true, $api_codes_class_name::OK, $message_or_api_code);
 	}
 
 	/**
@@ -138,7 +138,8 @@ class BuildResponseTest extends Base\ResponseBuilderTestCaseBase
 	public function testMake_CustomMessageAndWrongCode()
 	{
 		$api_code = [];    // invalid
-		$this->callMakeMethod($api_code, 'message');
+		/** @noinspection PhpParamsInspection */
+		$this->callMakeMethod(true, $api_code, 'message');
 	}
 
 	/**
@@ -149,7 +150,7 @@ class BuildResponseTest extends Base\ResponseBuilderTestCaseBase
 	public function testMake_CustomMessageAndCodeOutOfRange()
 	{
 		$api_code = $this->max_allowed_code + 1;    // invalid
-		$this->callMakeMethod($api_code, 'message');
+		$this->callMakeMethod(true, $api_code, 'message');
 	}
 
 }