+++ tests/SuccessTest.php
@@ -2,6 +2,7 @@
 
 namespace MarcinOrlowski\ResponseBuilder\Tests;
 
+use MarcinOrlowski\ResponseBuilder\ApiCodeBase;
 use MarcinOrlowski\ResponseBuilder\ResponseBuilder;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
@@ -25,10 +26,65 @@ class SuccessTest extends Base\ResponseBuilderTestCaseBase
 	public function testSuccess()
 	{
 		$this->response = ResponseBuilder::success();
-		$j = $this->getResponseSuccessObject();
+		$j = $this->getResponseSuccessObject(ApiCodeBase::OK);
+
+		$this->assertNull($j->data);
+		$this->assertEquals(\Lang::get(ApiCodeBase::getMapping(ApiCodeBase::OK)), $j->message);
+	}
+
+	/**
+	 * Tests success() with custom API code no custom message
+	 *
+	 * @return void
+	 */
+	public function testSuccess_ApiCode_NoCustomMessage()
+	{
+		\Config::set('response_builder.map', []);
+		$api_code = mt_rand($this->min_allowed_code, $this->max_allowed_code);
+
+		$this->response = ResponseBuilder::success(null, $api_code);
+		$j = $this->getResponseSuccessObject($api_code);
 
 		$this->assertNull($j->data);
 	}
+
+	/**
+	 * Tests success() with custom API code and no custom message mapping
+	 *
+	 * @return void
+	 */
+	public function testSuccess_ApiCode_CustomMessage()
+	{
+		$this->response = ResponseBuilder::success(null, $this->random_api_code);
+		$j = $this->getResponseSuccessObject($this->random_api_code);
+
+		$this->assertNull($j->data);
+	}
+
+
+	/**
+	 * Tests success() with custom API code and custom message
+	 *
+	 * @return void
+	 */
+	public function testSuccess_ApiCode_CustomMessageLang()
+	{
+		// for simplicity let's reuse existing message that is using placeholder
+		\Config::set('response_builder.map', [
+			$this->random_api_code => ApiCodeBase::getMapping(ApiCodeBase::NO_ERROR_MESSAGE)
+		]);
+
+		$lang_args = [
+			'api_code' => $this->getRandomString('foo'),
+		];
+
+		$this->response = ResponseBuilder::success(null, $this->random_api_code, $lang_args);
+		$expected_message = \Lang::get(ApiCodeBase::getMapping($this->random_api_code), $lang_args);
+		$j = $this->getResponseSuccessObject($this->random_api_code, null, $expected_message);
+
+		$this->assertNull($j->data);
+	}
+
 
 	/**
 	 * Checks success() with valid payload and HTTP code
@@ -50,7 +106,7 @@ class SuccessTest extends Base\ResponseBuilderTestCaseBase
 
 		foreach ($payloads as $payload) {
 			foreach ($http_codes as $http_code_expect => $http_code_send) {
-				$this->response = ResponseBuilder::success($payload, $http_code_send);
+				$this->response = ResponseBuilder::success($payload, null, [], $http_code_send);
 
 				$j = $this->getResponseSuccessObject($api_codes_class_name::OK, $http_code_expect);
 
@@ -86,7 +142,7 @@ class SuccessTest extends Base\ResponseBuilderTestCaseBase
 	 *
 	 * @expectedException \InvalidArgumentException
 	 */
-	public function testSuccessErrorCodeMustBeInt()
+	public function testSuccess_ApiCodeMustBeInt()
 	{
 		ResponseBuilder::success(null, 'foo');
 	}
@@ -98,7 +154,7 @@ class SuccessTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testSuccess_HttpCodeNull()
 	{
-		$this->response = ResponseBuilder::successWithHttpCode(null);
+		ResponseBuilder::successWithHttpCode(null);
 	}
 
 	/**