+++ tests/TestModel.php
@@ -2,9 +2,6 @@
 
 namespace MarcinOrlowski\ResponseBuilder\Tests;
 
-use MarcinOrlowski\ResponseBuilder\ResponseBuilder;
-use Symfony\Component\HttpFoundation\Response as HttpResponse;
-
 /**
  * Laravel API Response Builder
  *