+++ CHANGES.md
@@ -1,9 +1,15 @@
 # API Response Builder for Laravel 5 #
 
+
+See [compatibility docs](docs/compatibility.md) for details about backward compatibility!
+
+
 ## CHANGE LOG ##
 
 * dev
-   * Corrected default config file having `use` entries
+   * **PARTIAL BACKWARD INCOMPATIBILITY**
+   * [RB-17] `success()` now allows to return api code as well
+   * Corrected default config file containing faulty and unneeded `use` entries
    * [RB-20] Renamed ErrorCode to ApiCodeBase
    * ApiCodeBase's `getMinCode()` and `getMaxCode()` are now `public`
    * Improved error messages to be even more informative
@@ -11,6 +17,7 @@
    * Renamed error_code param to api_code in all method signatures
    * `:api_code` is now code placeholder in strings (`:error_code` is still supported for compatibility, but obsolete)
    * Default HTTP codes are now declared as constants `DEFAULT_HTTP_CODE_xxx` if you need to know them
+   * ApiCodeBase::getMap() now ensures `map` config entry of expected `array` type
 
 * v2.2.1 (2017-02-20)
    * Documentation split into separate files